==> HaberDuygusuTahmini/1-)VerilerinSeleniumİleÇekilmesi.ipynb <==
from selenium import webdriver
import time
import pandas as pd

driver = webdriver.Chrome(executable_path=r'C:\Users\Ozgur Kucet\Desktop\chromedriver.exe')
driver.get('http://www.thegeyik.com/category/kisisel-gelisim/page/2/')


# 1.Veriler
#https://onedio.com/etiket/kizgin-/5731f2acd6db161f30e8d6fa     0

# 2.Veriler
#https://onedio.com/etiket/cok-aci-/562e40f6df878a1246186f07    0

# 3.Veriler
#https://onedio.com/etiket/cok-iyi-/562e40f6df878a1246186f02    1

# 4.Veriler
# https://onedio.com/etiket/helal-olsun-/562e40f6df878a1246186f03    1

#5.Veriler
#http://www.thegeyik.com/category/kesin-bilgi/    1

#6.Veriler 
#http://www.thegeyik.com/category/kisisel-gelisim/page/2/

#7.Veriler
#http://www.thegeyik.com/category/kultur/bilim/page/2/

#8.Veriler
#http://www.thegeyik.com/category/kultur/page/2/

#9.Veriler
#http://www.thegeyik.com/category/hayvanlar/page/2/

#10.Veriler
#http://www.thegeyik.com/category/sosyal-medya/page/2/

#11.Veriler
#http://www.thegeyik.com/category/komik/page/2/

#12.Veriler
#http://www.thegeyik.com/category/seyahat/page/2/

#13.Veriler
#http://www.thegeyik.com/category/fotograf-yasam/page/2/

time.sleep(3)
"""
sayfa = driver.find_element_by_xpath("/html/body/div[3]/div/div/div[1]/div[3]/ul/li[2]")
#print(sayfa.href)
sayfa.click()
"""
sayfa = driver.find_element_by_xpath("//*/div/div/div/div/div[1]/div/ul/li[17]/a/div/div[2]/div/h2")
print(sayfa.text)

#//*/div/div/div/div/div[1]/div/div/div/div/a[2]
#//*/div/div/div/div/div[1]/div/div/div/div/a[3]

#//*/div/div/div/div/div[1]/div/ul/li[17]/a/div/div[2]/div/h2
#//*/div/div/div/div/div[1]/div/ul/li[18]/a/div/div[2]/div/h2
sayfanınNumarası = 3
cümleler = []

while(True):    
    sayi = 1
    while(True):
        try:
            cümleler.append(driver.find_element_by_xpath("//*/div/div/div/div/div[1]/div/ul/li[" +str(sayi)+ "]/a/div/div[2]/div/h2").text)
            sayi = sayi+1
        except:
            break

    try:
        sayfa = driver.find_element_by_xpath("//*/div/div/div/div/div[1]/div/div/div/div/a[" +str(sayfanınNumarası) + "]")        
        sayfa.click()
        sayfanınNumarası = sayfanınNumarası+1 
        time.sleep(3)
    except:
        break


print(len(cümleler))

import pandas as pd
d = {'Text': cümleler, 'Sentiment': 1}
df = pd.DataFrame(data=d)
df

#df.to_csv("State13.csv",index = False)
