==> HaberDuygusuTahmini/2-)VerilerinBeatifulSoupİleÇekilmesi.ipynb <==
# -*- coding: utf-8 -*-
import time
import requests
from bs4 import BeautifulSoup

time.sleep(1)
liste = []
for i in range(1,10): #onedio dan kaç sayfa çekilmesini istiyorsak o kadar sayfa yazıyoruz.
    response = requests.get("https://onedio.com/goygoy/" +str(i)) #Şuan goygoydan 10 sayfa alıyorum farklı başlıkları almak için linkini buraya giriyorum
    time.sleep(1)
    soup = BeautifulSoup(response.content,"lxml")#html parser yerine hızlı olması için lxml kullandım.
    haberler = soup.find_all("article",attrs = {"class":"o-card medium row o-article"})#haberlerin bulunduğu class ı yazdım.
    for haber in haberler: #Tüm haber başlıklarını listeye yazdırdım.
        haberbaslık = haber.find_all("h3")
        liste.append(haberbaslık[0].text)

print(len(liste))#kaç başlık alındığı

import pandas as pd #Verilerimizi Bir pandas dataframe haline çeviriyoruz.
d = {'Text': liste, 'Sentiment': 1}
df = pd.DataFrame(data=d)
df

df.to_csv("onedioGoygoy2.csv",index = False) #Son olarakta csv olarak kaydediyoruz.
