==> HaberDuygusuTahmini/3-)VerilerinBeatifulSoupİleÇekilmesi.ipynb <==
# -*- coding: utf-8 -*-
import time
import requests
from bs4 import BeautifulSoup

time.sleep(1)
liste = []

#Aynı işlemleri thegeyik web sitesi için yaptım bu sitede genel olarak iyi haberler var. 

for i in range(1,40):
    response = requests.get("http://www.thegeyik.com/category/kultur/page/" +str(i)+ "/")
    time.sleep(1)
    soup = BeautifulSoup(response.content,"lxml")
    haberler = soup.find_all("li",attrs = {"class":"mvp-blog-story-wrap left relative infinite-post"})
    for haber in haberler:
        haberbaslık = haber.find_all("h2")
        liste.append(haberbaslık[0].text)


print(len(liste))

import pandas as pd
d = {'Text': liste, 'Sentiment': 1}
df = pd.DataFrame(data=d)
df

df.to_csv("thegeyikKültür.csv",index = False)
