==> HaberDuygusuTahmini/4-)DuyguAnaliziEnUygunÖğrenmeninBulunması.ipynb <==
# Haber Tahmini İçin En Uygun Makine Öğrenmesi Algoritmasının Bulunması

from textblob import TextBlob
from sklearn import model_selection, preprocessing ,linear_model,naive_bayes, metrics
from sklearn.feature_extraction.text import TfidfVectorizer ,CountVectorizer
from sklearn import decomposition, ensemble

import pandas ,xgboost, numpy, textblob, string
from keras.preprocessing import text, sequence
from keras import layers, models, optimizers

import pandas as pd #Verileri okumak
data1 = pd.read_csv('onedioÇokAcı.csv')
data2 = pd.read_csv('onedioKızgın2.csv')
data3 = pd.read_csv('thegeyikHayvanlar.csv')
data4 = pd.read_csv('thegeyikKomik.csv')
data5 = pd.read_csv('thegeyikSeyehat.csv')
data6 = pd.read_csv('onedioGoygoy.csv')
data7 = pd.read_csv('onedioGoygoy2.csv')
data=pd.concat([data1,data2,data3,data4,data5,data6,data7], ignore_index = True) #Verileri birleştirmek
data

## Sentiment Analizi Ve Sınıflandırma Modelleri

data["Sentiment"].replace(0,value = "negatif" ,inplace = True) #0 lara negatif 
data["Sentiment"].replace(1,value = "pozitif" ,inplace = True) #1 lere pozitif

data

data.groupby("Sentiment").count()

df = pd.DataFrame() 
df["text"] = data["Text"]
df["label"] = data ["Sentiment"]
df.head()

## Metin Ön İşleme

#buyuk-kucuk donusumu
df['text'] = df['text'].apply(lambda x: " ".join(x.lower() for x in x.split()))
#noktalama işaretleri
df['text'] = df['text'].str.replace('[^\w\s]','')
#sayılar
df['text'] = df['text'].str.replace('\d','')

pd.Series(" ".join(df["text"]).split()).value_counts() 

#stopwords
import nltk
#nltk.download('stopwords')
from nltk.corpus import stopwords
sw = stopwords.words('Turkish')
df['text'] = df['text'].apply(lambda x: " ".join(x for x in x.split() if x not in sw))

print(sw) #Stopwordsler aslında dildeki sık kullanılan kelimeler bunları veri setimizde istemiyoruz.

# Burada kelime köklerini 
from snowballstemmer import TurkishStemmer
turkStem=TurkishStemmer()
#nltk.download('wordnet')
from textblob import Word
 #nltk.download('wordnet')
df['text'] = df['text'].apply(lambda x: " ".join([turkStem.stemWord(word) for word in x.split()])) #kelimelerin köklerine indirgedik. Bu işlem veri işleme için önemli bir adım.

#seyreklerin silinmesi
sil = pd.Series(' '.join(df['text']).split()).value_counts()[-1000:]
df['text'] = df['text'].apply(lambda x: " ".join(x for x in x.split() if x not in sil))

df.to_csv("ÖnİşlenmişVeriler2.csv",index = False)

## Değişken Mühendisliği

Bir şekilde metinleri ölcülebilir metriklere dönüştürmek istiyoruz. 
- Count Vectors: Basit sayma işlemleri
- TF-IDF Vectors(words,chracters,n-grams)
- TF(t) = (Bir t teriminin bir dökümanda gözlenme frekansı / dökümandaki toplam terim sayısı)
- IDF(t) = log_e(Toplam döküman sayısı içinde t terimi olan belge sayısı)

## Test-Train

train_x ,test_x,train_y,test_y = model_selection.train_test_split(df["text"],
                                                                 df["label"])

encoder = preprocessing.LabelEncoder()

train_y = encoder.fit_transform(train_y)
test_y = encoder.fit_transform(test_y)

## Count Vector

vectorizer = CountVectorizer()
vectorizer.fit(train_x)

X_train_count = vectorizer.transform(train_x)
X_test_count = vectorizer.transform(test_x)

vectorizer.get_feature_names()[10:20] 

X_train_count.toarray() 

## TDF-IDF(words)

tf_idf_word_vectorizer = TfidfVectorizer()
tf_idf_word_vectorizer.fit(train_x)

X_train_tf_idf_word = tf_idf_word_vectorizer.transform(train_x)
X_test_tf_idf_word = tf_idf_word_vectorizer.transform(test_x)

tf_idf_word_vectorizer.get_feature_names()[10:50] #içerisindeki verilere bakmak.

X_train_tf_idf_word.toarray()

## TDF-IDF(ngram)

tf_idf_ngram_vectorizer = TfidfVectorizer(ngram_range=(2,3))
tf_idf_ngram_vectorizer.fit(train_x)

X_train_tf_idf_ngram = tf_idf_ngram_vectorizer.transform(train_x)
X_test_tf_idf_ngram = tf_idf_ngram_vectorizer.transform(test_x)

## TDF-IDF(Characters)

tf_idf_chars_vectorizer = TfidfVectorizer(analyzer = "char", ngram_range=(2,3))
tf_idf_chars_vectorizer.fit(train_x)

X_train_tf_idf_chars = tf_idf_chars_vectorizer.transform(train_x)
X_test_tf_idf_chars = tf_idf_chars_vectorizer.transform(test_x)

## Makine Öğrenmesi İle Doğruluk Oranlarının Bulunması

## Logistic Regression

loj = linear_model.LogisticRegression()
loj_model = loj.fit(X_train_count, train_y)
accuracy = model_selection.cross_val_score(loj_model, 
                                           X_test_count, 
                                           test_y, 
                                           cv = 10).mean()

print("Count Vectors Doğruluk Oranı:", accuracy)

loj = linear_model.LogisticRegression()
loj_model = loj.fit(X_train_tf_idf_word,train_y)
accuracy = model_selection.cross_val_score(loj_model, 
                                           X_test_tf_idf_word, 
                                           test_y, 
                                           cv = 10).mean()

print("TF-IDF Doğruluk Oranı:", accuracy)

loj = linear_model.LogisticRegression()
loj_model = loj.fit(X_train_tf_idf_ngram,train_y)
accuracy = model_selection.cross_val_score(loj_model, 
                                           X_test_tf_idf_ngram, 
                                           test_y, 
                                           cv = 10).mean()

print("N-GRAM TF-IDF Doğruluk Oranı:", accuracy)

loj = linear_model.LogisticRegression()
loj_model = loj.fit(X_train_tf_idf_chars,train_y)
accuracy = model_selection.cross_val_score(loj_model, 
                                           X_test_tf_idf_chars, 
                                           test_y, 
                                           cv = 10).mean()

print("CHARLEVEL Doğruluk Oranı:", accuracy)

## Naive Bayes

nb = naive_bayes.MultinomialNB()
nb_model = nb.fit(X_train_tf_idf_word,train_y)
accuracy = model_selection.cross_val_score(nb_model, 
                                           X_test_tf_idf_word, 
                                           test_y, 
                                           cv = 10).mean()

print("Word-Level TF-IDF Doğruluk Oranı:", accuracy)

nb = naive_bayes.MultinomialNB()
nb_model = nb.fit(X_train_tf_idf_ngram,train_y)
accuracy = model_selection.cross_val_score(nb_model, 
                                           X_test_tf_idf_ngram, 
                                           test_y, 
                                           cv = 10).mean()

print("N-GRAM TF-IDF Doğruluk Oranı:", accuracy)

nb = naive_bayes.MultinomialNB()
nb_model = nb.fit(X_train_tf_idf_chars,train_y)
accuracy = model_selection.cross_val_score(nb_model, 
                                           X_test_tf_idf_chars, 
                                           test_y, 
                                           cv = 10).mean()

print("CHARLEVEL Doğruluk Oranı:", accuracy)

## Random Forest 

rf = ensemble.RandomForestClassifier()
rf_model = rf.fit(X_train_count,train_y)
accuracy = model_selection.cross_val_score(rf_model, 
                                           X_test_count, 
                                           test_y, 
                                           cv = 10).mean()

print("Count Vectors Doğruluk Oranı:", accuracy)

rf = ensemble.RandomForestClassifier()
rf_model = rf.fit(X_train_tf_idf_word,train_y)
accuracy = model_selection.cross_val_score(rf_model, 
                                           X_test_tf_idf_word, 
                                           test_y, 
                                           cv = 10).mean()

print("Word-Level TF-IDF Doğruluk Oranı:", accuracy)

rf = ensemble.RandomForestClassifier()
rf_model = loj.fit(X_train_tf_idf_ngram,train_y)
accuracy = model_selection.cross_val_score(rf_model, 
                                           X_test_tf_idf_ngram, 
                                           test_y, 
                                           cv = 10).mean()

print("N-GRAM TF-IDF Doğruluk Oranı:", accuracy)

rf = ensemble.RandomForestClassifier()
rf_model = loj.fit(X_train_tf_idf_chars,train_y)
accuracy = model_selection.cross_val_score(rf_model, 
                                           X_test_tf_idf_chars, 
                                           test_y, 
                                           cv = 10).mean()

print("CHARLEVEL Doğruluk Oranı:", accuracy)

## XGboost

xgb = xgboost.XGBClassifier()
xgb_model = xgb.fit(X_train_count,train_y)
accuracy = model_selection.cross_val_score(xgb_model, 
                                           X_test_count, 
                                           test_y, 
                                           cv = 10).mean()

print("Count Vectors Doğruluk Oranı:", accuracy)

xgb = xgboost.XGBClassifier()
xgb_model = xgb.fit(X_train_tf_idf_word,train_y)
accuracy = model_selection.cross_val_score(xgb_model, 
                                           X_test_tf_idf_word, 
                                           test_y, 
                                           cv = 10).mean()

print("Word-Level TF-IDF Doğruluk Oranı:", accuracy)

xgb = xgboost.XGBClassifier()
xgb_model = xgb.fit(X_train_tf_idf_chars,train_y)
accuracy = model_selection.cross_val_score(xgb_model, 
                                           X_test_tf_idf_chars, 
                                           test_y, 
                                           cv = 10).mean()

print("CHARLEVEL Doğruluk Oranı:", accuracy)

### En iyi Doğruluk oranını veren 0.8856304969513842 doğruluk oranı Count Vectors İle LogisticRegression algoritmasında oldu.
